Caching main index views
This commit adds caching for the main index views (without filters) to
speed up the "typical" use case where many users are viewing the latest
topics.

To still support logged in user / team features (icons/badges), all of
these features were extracted into an eager loaded turbo frame, so that
even if we generate the cache with a logged in user, we do not include
any user specific details in the display.