Fix caching bug on topic index
We included the user-specific async data retrieval in the cached block,
but only if we had a signed in user.

This worked perfectly if the page first was cached by a signed in user,
but failed if it was first cached by a guest, as after that signed in
users didn't get their additional information rendered.

This commit fixes the problem by moving the user specific part outside
of the cached part.